hide skill selection until family is selected
