+++ apps/web/src/components/SkillDialog/types.ts
@@ -1,1 +1,10 @@
+import { Scalars, SkillCategory } from "@gc-digital-talent/graphql";
+
 export type SkillDialogContext = "experience" | "library" | "showcase";
+
+export interface FormValues {
+  category?: SkillCategory | "all" | "";
+  family?: Scalars["ID"];
+  skill?: Scalars["ID"];
+  details?: string;
+}
